{{ cookiecutter.project_name }}/src/pipeline/requirements.txt
kfp==2.3.0
google-cloud-aiplatform==1.36.2
google-cloud-pipeline-components==2.5.0
